Criacao de construtores e metodos utilitarios.
